+++ src/main/java/balancer/storage/TransactionComparator.java
@@ -8,6 +8,6 @@ import java.util.Comparator;
 public class TransactionComparator implements Comparator<Transaction> {
     @Override
     public int compare(Transaction t1, Transaction t2) {
-        return Float.compare(t1.getAmount(), t2.getAmount());
+        return Double.compare(t1.getAmount(), t2.getAmount());
     }
 }
